build/toolchains: restore workspace clippy lint policy
Reintroduces the deny/allow set from the pre-buck2 [workspace.lints.clippy]
table, applied via the rust toolchain's deny_lints/allow_lints. Same
treatment for tests, fuzz, loom, and production builds.

Also applies fixes so the now enabled lint actually passes.

+++ lib/ecma-119/src/raw/mod.rs
@@ -19,4 +19,4 @@ pub use volume::*;
 
 pub const SECTOR_SIZE: usize = 2048;
 /// El Torito uses 512-byte "virtual" sectors for boot image sector counts.
-pub const VIRTUAL_SECTOR_SIZE: usize = 512;
+pub const VIRTUAL_SECTOR_SIZE: u32 = 512;
